+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Tab component focus fixed - focus ring only show on keyboard interactions
 - Prevent www theme examples from bleeding into global scope
 - Fix slider overflow issue in Windows/Edge
 - Fix www sidebar nav button appearance in Windows/Edge